+++ conformance/utils/features/features.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package suite
+package features
 
 import "k8s.io/apimachinery/pkg/util/sets"
